Permissions Management Rework; stopForeground() on ScanService.onDestroy
Totally reworked how permissions are requested. It's very thoroughly commented in-code, but briefly:

Checks for each permission we need. If that permission is currently denied, we request it.

Not very different to the end-user (still just requests location access), but, to my knowledge, much more appropriate for API 28/29.

Also, by adding stopForeground(true) to ScanService.onDestroy, not only is the notification removed, but this should stop Android from notifying user "[app] got your location in background" when not scanning.

Tested and working on my S10 Plus with Android 10.